Reapply "Bug fixes and Readme app"
This reverts commit 84acecef3b1354cd9bfda58916b23a61bc680412.

+++ tests/utils/mocks.c
@@ -162,7 +162,7 @@ void* kalloc(int size){
 }
 
 int kcalloc(int size){
-    return kalloc(size);
+    return (int)kalloc(size);
 }
 
 #endif
